Add validation callback mechanism for ffmpeg and retry when validation fails
